pyproject.toml
[project]
name = "nyc-taxi-pipeline"
version = "1.0.0"
description = "End-to-end ETL pipeline for NYC TLC trip data analysis"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.1",
    "loguru>=0.7.2",
    "duckdb>=0.10.0",
    "pandas>=2.2.0",
    "pyarrow>=15.0.0",
    "httpx>=0.26.0",
    "aiofiles>=23.2.1",
    "prefect>=2.14.21",
    "rich>=13.7.0",
    "click>=8.1.7",
    "python-dateutil>=2.8.2",
]

[dependency-groups]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "ruff>=0.13.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
asyncio_mode = "auto"

